+++ src/main/java/org/truffleruby/language/objects/CheckIVarNameNode.java
@@ -13,7 +13,6 @@ import com.oracle.truffle.api.profiles.BranchProfile;
 import org.truffleruby.core.symbol.RubySymbol;
 import org.truffleruby.core.symbol.SymbolTable;
 import org.truffleruby.language.RubyBaseNode;
-import org.truffleruby.language.RubyDynamicObject;
 import org.truffleruby.language.control.RaiseException;
 import org.truffleruby.parser.IdentifierType;
 import org.truffleruby.parser.Identifiers;
@@ -31,10 +30,10 @@ public abstract class CheckIVarNameNode extends RubyBaseNode {
 
     /** Pass both the j.l.String name and the original name, the original name can be faster to check and the j.l.String
      * name is needed by all callers so it is better for footprint that callers convert to j.l.String */
-    public abstract void execute(RubyDynamicObject object, String name, Object originalName);
+    public abstract void execute(Object object, String name, Object originalName);
 
     @Specialization
-    protected void checkSymbol(RubyDynamicObject object, String name, RubySymbol originalName,
+    protected void checkSymbol(Object object, String name, RubySymbol originalName,
             @Cached BranchProfile errorProfile) {
         if (originalName.getType() != IdentifierType.INSTANCE) {
             errorProfile.enter();
@@ -48,12 +47,12 @@ public abstract class CheckIVarNameNode extends RubyBaseNode {
     @Specialization(
             guards = { "name == cachedName", "isValidInstanceVariableName(cachedName)", "!isRubySymbol(originalName)" },
             limit = "getDynamicObjectCacheLimit()")
-    protected void cached(RubyDynamicObject object, String name, Object originalName,
+    protected void cached(Object object, String name, Object originalName,
             @Cached("name") String cachedName) {
     }
 
     @Specialization(replaces = "cached", guards = "!isRubySymbol(originalName)")
-    protected void uncached(RubyDynamicObject object, String name, Object originalName) {
+    protected void uncached(Object object, String name, Object originalName) {
         SymbolTable.checkInstanceVariableName(getContext(), name, object, this);
     }
 }